Create Simple Maze Text Adventure - Updated
I have completed the text adventure task that Zahra had us do, and I also updated it to include a while loop. This means that the player can play over and over again for as many times as they desire as long as they input 'y' after every playthrough. However, if they input 'n' at the end of the playthrough, it will stop the while loop, therefore ending the program.

:)